add floating support (tested)

+++ src/Calculator.Core/Calculation/PolishCalculate.cs
@@ -7,7 +7,7 @@ namespace Calculator.Core
 
     public interface ICalculate// needed only for mocking PolishCalculate
     {
-        int Calculate();
+        float Calculate();
     }
 
     public class PolishCalculate:ICalculate
@@ -17,11 +17,11 @@ namespace Calculator.Core
 
         public PolishCalculate(IEnumerable<Token> tokens) => this.m_tokens = tokens;
 
-        public int Calculate()
+        public float Calculate()
         {
             if (!this.m_tokens.Any())
                 throw new NoTokensException();
-            Stack<int> stack = new Stack<int>();
+            Stack<float> stack = new Stack<float>();
 
             foreach (var token in this.m_tokens)
             {
@@ -31,14 +31,14 @@ namespace Calculator.Core
             return stack.Pop();
         }
 
-        private void Calculate(Stack<int> stack, Token token)
+        private void Calculate(Stack<float> stack, Token token)
         {
             if (token is OperatorToken)
             {
                 if (stack.Count < 2)
                     throw new NotValidOperationException("Not enuoth numeric");
-                int left = stack.Pop();
-                int right = stack.Pop();
+                float left = stack.Pop();
+                float right = stack.Pop();
 
                 switch (((OperatorToken)token)._value)
                 {
